Revert the morphable options idea

+++ plugin/blink/parameters/option_data.hpp
@@ -10,13 +10,13 @@ class OptionIndexData
 public:
 
 	const blink_OptionData* const data;
-	const float default_value;
-	const float value;
+	const int32_t default_value;
+	const int32_t value;
 	const OptionParameter& option;
 
 	OptionIndexData(const OptionParameter& option_, const blink_ParameterData* param_data, blink_Index index)
 		: data { param_data ? &param_data[index].option : nullptr }
-		, default_value { float(option_.default_index) }
+		, default_value { option_.default_index }
 		, value { data && data->data.count > 0 ? data->data.points[0].y : default_value }
 		, option { option_ }
 	{
